chore: simplify slices
