+++ ui/src/slices/globalStatus.ts
@@ -1,5 +1,9 @@
-import { Alert } from './globalStatus.type'
-import { createSlice } from '@reduxjs/toolkit'
+import { PayloadAction, createSlice } from '@reduxjs/toolkit'
+
+export interface Alert {
+  type: 'success' | 'warning' | 'error'
+  message: string
+}
 
 const initialState: {
   alert: Alert
@@ -18,13 +22,13 @@ const globalStatusSlice = createSlice({
   name: 'globalStatus',
   initialState,
   reducers: {
-    setAlert(state, action) {
+    setAlert(state, action: PayloadAction<Alert>) {
       state.alert = action.payload
     },
-    setAlertOpen(state, action) {
+    setAlertOpen(state, action: PayloadAction<boolean>) {
       state.alertOpen = action.payload
     },
-    setSearchModalOpen(state, action) {
+    setSearchModalOpen(state, action: PayloadAction<boolean>) {
       state.searchModalOpen = action.payload
     },
   },
