--- ui/src/slices/globalStatus.type.ts
@@ -1,4 +1,0 @@
-export interface Alert {
-  type: 'success' | 'error'
-  message: string
-}
