--- ui/src/slices/navigation.type.ts
@@ -1,8 +1,0 @@
-import { PayloadAction } from '@reduxjs/toolkit'
-
-export interface NavigationBreadCrumbProps {
-  name: string
-  path?: string
-}
-
-export type NavigateAction = PayloadAction<string>
